Implement generic pipeline caching
