+++ node-graph/nodes/gstd/Cargo.toml
@@ -61,6 +61,7 @@ reqwest = { workspace = true }
 image = { workspace = true }
 base64 = { workspace = true }
 wgpu = { workspace = true }
+bytemuck = { workspace = true }
 
 # Optional local dependencies
 graphene-canvas-utils = { workspace = true, optional = true }
