+++ node-graph/nodes/gstd/src/render_background_checker_rect.wgsl
