+++ node-graph/nodes/gstd/src/render_background_fullscreen.wgsl
