+++ node-graph/nodes/gstd/src/render_pixel_preview.wgsl
